refactor: refactor reg-form
