+++ src/features/auth/components/reg-form.tsx
@@ -6,22 +6,15 @@ import {
   loginUser,
   sanitizeCustomerDraft,
 } from '../services/auth-service'
-import {
-  isValidEmail,
-  isValidPassword,
-  isValidName,
-  isValidBirthDate,
-  isValidStreet,
-  isValidCity,
-  isValidPostalCode,
-  isValidCountry,
-} from '../../../utils/validators'
 import type {
   CustomerDraftPayload,
   FormDataType,
 } from '../../../types/customer'
-
-const validCountries = ['United States', 'Canada']
+import {
+  validate,
+  initialForm,
+  validCountries,
+} from '../../../utils/form-utils'
 
 const countryNameToCode: Record<string, string> = {
   Canada: 'CA',
@@ -30,22 +23,6 @@ const countryNameToCode: Record<string, string> = {
 
 export const RegForm = () => {
   const dispatch = useAppDispatch()
-
-  const initialForm: FormDataType = {
-    email: '',
-    password: '',
-    firstName: '',
-    lastName: '',
-    birthDate: '',
-    street: '',
-    city: '',
-    postalCode: '',
-    country: '',
-    billingStreet: '',
-    billingCity: '',
-    billingPostalCode: '',
-    billingCountry: '',
-  }
 
   const [formData, setFormData] = useState<FormDataType>(initialForm)
   const [errors, setErrors] = useState<{ [key: string]: string }>({})
@@ -63,48 +40,10 @@ export const RegForm = () => {
     const { name, value } = e.target
     const updatedForm = { ...formData, [name]: value }
     setFormData(updatedForm)
-    if (hasSubmitted) validate(updatedForm)
-  }
-
-  const validateBillingFields = (data: FormDataType) => {
-    const billingErrors: { [key: string]: string } = {}
-    if (!isValidStreet(data.billingStreet))
-      billingErrors.billingStreet = 'Billing street cannot be empty'
-    if (!isValidCity(data.billingCity))
-      billingErrors.billingCity = 'Invalid billing city'
-    if (!isValidPostalCode(data.billingPostalCode, data.billingCountry))
-      billingErrors.billingPostalCode = 'Invalid billing postal code'
-    if (!isValidCountry(data.billingCountry, validCountries))
-      billingErrors.billingCountry = 'Select a valid billing country'
-    return billingErrors
-  }
-
-  const validate = (data: FormDataType) => {
-    const newErrors: { [key: string]: string } = {}
-
-    if (!isValidEmail(data.email)) newErrors.email = 'Invalid email format'
-    if (!isValidPassword(data.password))
-      newErrors.password =
-        'Password must be at least 8 characters, include upper/lowercase, number and one special character'
-    if (!isValidName(data.firstName))
-      newErrors.firstName = 'Name shouldn`t include digits'
-    if (!isValidName(data.lastName))
-      newErrors.lastName = 'Name shouldn`t include digits'
-    if (!isValidBirthDate(data.birthDate))
-      newErrors.birthDate = 'You must be at least 13 years old'
-    if (!isValidStreet(data.street)) newErrors.street = 'Street cannot be empty'
-    if (!isValidCity(data.city)) newErrors.city = 'Invalid city'
-    if (!isValidPostalCode(data.postalCode, data.country))
-      newErrors.postalCode = 'Invalid postal code'
-    if (!isValidCountry(data.country, validCountries))
-      newErrors.country = 'Select a valid country'
-
-    if (!useSameAddress) {
-      Object.assign(newErrors, validateBillingFields(data))
+    if (hasSubmitted) {
+      const validationErrors = validate(updatedForm, useSameAddress)
+      setErrors(validationErrors)
     }
-
-    setErrors(newErrors)
-    return Object.keys(newErrors).length === 0
   }
 
   useEffect(() => {
@@ -136,7 +75,9 @@ export const RegForm = () => {
     setMessage('')
     setErrorMessage('')
 
-    if (!validate(formData)) return
+    const validationErrors = validate(formData, useSameAddress)
+    setErrors(validationErrors)
+    if (Object.keys(validationErrors).length > 0) return
 
     const addresses = [
       {
